Shared workflow and project build update (#25)
* readme: Add build badge

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>

* testing: Place reusable test notebook workflow for other repos to use

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>

* ci: Add dependabot config

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>

* build update: pyproject.toml and tox

The python build is updated to use pyproject.toml instead of the old
setup.py build. tox is also added to control the test environments and
add linting commands. With the new build, the code is moved into the
standard src and tests folders. The ci build is updated to use the new
tox setup including running lint jobs.

The code is updated to pass the lint checks and python typing is added.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>

---------

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>