Don't pass documents to completion LLMs
We update create_stuff_documents_chain to assume client-side prompt
formatting for completion LLMs since LangChain will covert the prompt
value to a string. Thus we don't pass documents to the model object.

For chat completion LLMs we pass documents to the model since the
messages (and documents) will be passed to the server for server-side
prompt formatting.

Note that some chat completion APIs don't support documents in the API.
For example Ollama. So we add an option to instead add the documents to
the messages as document role messages. The Granite templates for Ollama
understand document role messages when formatting the prompt.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>